ext/Pod-Html/t/*.t: Have each file run in tempdir
Until now, the tests for this library ran inside the ext/Pod-Html
directory in the core distribution.  Since the tests necessarily create
files while the harness runs, there was a possibility for race
conditions, which meant that we had to have a workaround in t/harness to
have the tests in this directory not run in parallel.

This commit restructures the testing process so that each test program
gets its own tempdir in which to create files and directories.

The test architecture is now found in t/lib/Testing.pm.
This module will export two functions on demand that are used in the
individual test files.  They guarantee that all files created when
running a test program will be created in a temporary directory rather
than underneath the core distribution.

t/pod2html-lib.pl is now superseded by t/lib/Testing.pm and can be
removed.  The workaround has been removed from t/harness.

Change directory name from testdir/ to corpus/.  So as to more precisely
describe what the directory's purpose is.

Keep porting tests happy. Update MANIFEST and known_pod_issues database.

Acknowledge Craig Berry's recommendation in earlier branch

This should help in the resolution of
https://github.com/Perl/perl5/issues/12271.

Correct typo spotted by rjbs.