regcomp.c - zero the pRExC_state structure early in compilation
And initialize various members, especially RExC_mysv1 and RExC_mysv2
as early as possible under debugging.

Under -Uusedl we were ending up with the mysv1 and mysv2 slots set to
random pointers as the structure wasnt zeroed properly, and the
initialization for the temporary sv's wasn't performed correctly, which
then lead to segfaults.

Arguably we should have been using Zero() from the very beginning, as
we have had multiple bugs over the years where we added a member to
the structure but forgot to zero it at the start of compilation. This
change rectifies that and will prevent future oversights.

(cherry picked from commit ecbf63ae7dd6638fee8184d452ebacf312c3b535)