ParseXS: fix bool return code: RETVAL with magic
When generate_output() is asked to emit code which returns the value of
RETVAL when that value is boolean, then in the specific case of
$do_setmagic being true, it emits invalid code.

Now it so happens that for RETVAL, generate_output() is *never* called
with $do_setmagic set to true, so this is all a bit academic. But I'm
fixing it in case later refactoring tries to use this code path for
variables other than RETVAL. The fact that it can never happen is why
this commit doesn't include any tests.

The actual issue is that: while normally a bool return is simplified to
this single line of code:

    ST(0) = boolSV(RETVAL);

when magic needs to be set then the value needs to be saved to a temporary
value, such as:

    SV * RETVALSV;
    RETVALSV = boolSV(RETVAL);
    SvSETMAGIC(RETVALSV);
    ST(0) = RETVALSV;

but the code had incorrectly concluded that it could use RETVAL instead:

    RETVAL = boolSV(RETVAL);
    SvSETMAGIC(RETVAL);
    ST(0) = RETVAL;

Even though RETVAL is of type bool rather than SV*. With this fix,
RETVALSV is used.