ParseXS: expand PUSHp to { sv_setpvn_mg; PUSHs }
When ParseXS emits code which uses the TARG optimisation to push
a PV onto the stack as the return value, it currently uses one of two
approaches, depending on whether there is a length argument:

    1) PUSHp(val, len);
    2) sv_setpv(TARG, val); PUSHTARG;

This commit expands and simplifies the output for those macros. In
particular, separating out the TARG-setting and TARG-pushing operations
to make future refactoring easier.

For the first, this commit expands the PUSHp() macro so that the emitted
code becomes:

    sv_setpvn(TARG, val, len); SvSETMAGIC(TARG); PUSHs(TARG);

and then moves the magic setting into the function call:

    sv_setpvn_mg(TARG, val, len); PUSHs(TARG);

For the second, it expands the PUSHTARG:

    sv_setpv(TARG, val); SvSETMAGIC(TARG); PUSHs(TARG);

and again, then moves the magic setting into the function call:

    sv_setpv_mg(TARG, val); PUSHs(TARG);