ParseXS: refactor: TARG opt: do eval only once
For the TARG optimisation, the typemap parser has detected typemaps
of the form sv_setiv(), sv_setpvn() etc and split them out into their
components. These components will then be variable-expanded using eval,
then put back in a modified version of the original call.
So for example, for a typemap of

    sv_setiv($arg, (IV)$var);

we would do the equivalent of:

    $what = '(IV)$var';   # extracted out by typemap parser
    $what = $self->eval_output_typemap_code($what, {var => $var, ...});
    print "TARGi($what, 1);\n";

However for the sv_setpvn() case, there is an extra length arg.
Formerly this was variable-expanded using a *second* eval. This commit
combines both into a single eval, making the code simpler. So for this
typemap:

    sv_setpvn($arg, (char *)&$var, sizeof($var));

it now does the equivalent of:

    $what  = '(char *)&$var';  # extracted out by typemap parser
    $tsize = ', sizeof($var)'; # extracted out by typemap parser
    $what = $self->eval_output_typemap_code("$what$tsize", {var => $var, ...});
    print "sv_setpvn_mg(TARG, $what);\n";