ParseXS: refactor: pull out common XSprePUSH emit
The two branches of the TARG optimisation both emit a 'XSprePUSH;' line.
This commit hoists those two identical blocks of code to outside the
if/else.

For the TARGi/u/n case, as a side effect it changes the ordering of the
lines of C code from:

    XSprePUSH;           /* reset stack pointer to base of frame */
    TARGu((IV)RETVAL, 1);
    PUSHs(TARG);

to

    TARGu((IV)RETVAL, 1);
    XSprePUSH;           /* reset stack pointer to base of frame */
    PUSHs(TARG);

Which is functionally the same.

For the sv_setpv/sv_setpvn cases, there is no change in output (so this
commit actually makes things more consistent).