ParseXS: refactor: TARG opt: eval whole expression
Given a typemap like

    sv_setiv($arg, (IV)$var);

the TARG optimisation currently extracts out the '(IV)$var' part of it,
evals that as a double-quoted string in order to variable-expand '$var'
etc, then constructs a new expression using that result, e.g.
'TARGi((IV)RETVAL)'.

This commit changes it so that the eval expansion comes *after* the
reconstruction. So before:

    $what = '(IV)$var';
    $what = eval qq("$what");
    $result = "TARGi($what)";

and after:

    $what = '(IV)$var';
    $what = "TARGi($what)";
    $result = eval qq("$what");

The result should be the same, but it makes the code closer to how
RETVAL is handled in the rest of generate_output().