perly LSTOPSUB: prevent a double op free between parse stack and CV
The reeproducer resulted in the "block" OP being both on the parser
stack and attacked to the CV.  If an error occurred while parsing the
rest of the list operator clean up would release the OP as attached
to the CV, and the same OP on the parse stack, resulting in a double
free.

It's unclear to me whether bison is intended to support modifying
the parse stack entry like this, but it appears to work here.

Fixes #21724

(MANIFEST, perly.y and t/lib/croak/parser manually cherry picked from
commit 0d9e3bc6dbe88da0464ba4b77219ed5b7c276340; perly.act, perly.h and
perly.tab regenerated using regen_perly.pl with Cygwin bison 3.7.6)