doop.c Use new utf8_to_uv_or_die()
This replaces the older forms.  For two of the instances, this gives
equivalent functionality, where they croaked if an operand to tr/// was
malformed.  For the other instances, it didn't croak, but now does.
This makes the behavior consistent for tr///, as all the affected
instances are supposed to be equivalent.  For performance reasons, one
or another of the functions would be called, and depending on which one
the same input would croak or not.