fix refassign in non-void context
GH #22710

The experimental refassign feature, \@a = $aref, which allows an
array/hash etc to be aliased rather than copied, got broken in rvalue
non-void context by v5.39.5-56-g604c0355d2. So in something like

    my $x = (\@a = $aref);

the stack was underflowing. This is because previously pp_refassign()
would leave its RH argument on the stack when returning, while my commit
above made it always pop the argument off the stack.

This commit makes it only pop the arg in void context.